+++ hail/python/hail/utils/__init__.py
@@ -11,7 +11,7 @@ from .frozendict import frozendict
 from .java import error, warning, info, FatalError, HailUserError
 from .tutorial import get_1kg, get_hgdp, get_movie_lens
 from .deduplicate import deduplicate
-from .json import JSONEncoder
+from .jsonx import JSONEncoder
 
 __all__ = ['hadoop_open',
            'hadoop_copy',
