+++ hail/python/hail/utils/jsonx.py
